Added a Rust program to wrap bash scripts
This program will wrap any given script into an exectuable.
It's main purpose is to allow scripts to be able to be executed with
the SUID bit, as this is normally not not supported due to security reasons.

To make it quite secure, the rust executalbe clears all env vars and doesn't pass any argument. However it does parse /etc/environment, if the permissions are correct to give some basic default environment to scripts.