Setup and cleanup scripts don't require entering the root password
The required commands that had to be executed as root were splitted into
a separate file. Then these files are converted to a normal executable in
order to be able to set the SUID bit, so that they get elevated automatically
to root privileges.

Furthermore all setup related things are now nicely grouped under /opt/setup
and the home directory only contains symlinks.