Replaced Cheese by guvcview
As Cheese is currently bugged (see issue #39 and #38), we instead use
guvcview as a replacement. Kamoso was also tested as a replacement,
but crashed on certain computers (in our case on the aspire 7550g).
Cheese was also removed to avoid kids declaring the computer as defect,
because of a bugged webcam image.