Now enabling the heyalter service when executing the cleanup
Previously it was enabled at the setup stage, but as the Cleanup script
is always executed later on, we can better put it into the cleanup
to increase our chance of not triggering it before the kid gets it.

Furthermore the command to manually reenable the service was added
to the README, so people can even do it after the cleanup has happened.

Fixes #12